fix(browser): settle waits across navigation and shutdown
